refactor/#396 : 비즈니스 검증 Input Port의 Command에서 수행하도록 수정
